Create AuthValidation and move the logic from SkillValidation to AuthValidation
